fix: Try updating Podfile deployment target on prepare
