+++ lib/Api.js
@@ -34,7 +34,6 @@ const {
     PluginManager
 } = require('cordova-common');
 const util = require('util');
-const xcode = require('xcode');
 
 function setupEvents (externalEventEmitter) {
     if (externalEventEmitter) {
@@ -417,7 +416,7 @@ class Api {
                 projectFile.purgeProjectFileCache(this.locations.root);
 
                 return podfileFile.install(check_reqs.check_cocoapods)
-                    .then(() => this.setSwiftVersionForCocoaPodsLibraries(podsjsonFile));
+                    .then(() => podsjsonFile.setSwiftVersionForCocoaPodsLibraries(this.root));
             } else {
                 events.emit('verbose', 'Podfile unchanged, skipping `pod install`');
             }
@@ -506,7 +505,7 @@ class Api {
                 events.emit('verbose', 'Running `pod install` (to uninstall pods)');
 
                 return podfileFile.install(check_reqs.check_cocoapods)
-                    .then(() => this.setSwiftVersionForCocoaPodsLibraries(podsjsonFile));
+                    .then(() => podsjsonFile.setSwiftVersionForCocoaPodsLibraries(this.root));
             } else {
                 events.emit('verbose', 'Podfile unchanged, skipping `pod install`');
             }
@@ -520,43 +519,7 @@ class Api {
      * @param  {PodsJson}  podsjsonFile  A PodsJson instance that represents pods.json
      */
     setSwiftVersionForCocoaPodsLibraries (podsjsonFile) {
-        let __dirty = false;
-        return check_reqs.check_cocoapods().then(toolOptions => {
-            if (toolOptions.ignore) {
-                events.emit('verbose', '=== skip Swift Version Settings For Cocoapods Libraries');
-            } else {
-                const podPbxPath = path.join(this.root, 'Pods', 'Pods.xcodeproj', 'project.pbxproj');
-                const podXcodeproj = xcode.project(podPbxPath);
-                podXcodeproj.parseSync();
-                const podTargets = podXcodeproj.pbxNativeTargetSection();
-                const podConfigurationList = podXcodeproj.pbxXCConfigurationList();
-                const podConfigs = podXcodeproj.pbxXCBuildConfigurationSection();
-
-                const libraries = podsjsonFile.getLibraries();
-                Object.keys(libraries).forEach(key => {
-                    const podJson = libraries[key];
-                    const name = podJson.name;
-                    const swiftVersion = podJson['swift-version'];
-                    if (swiftVersion) {
-                        __dirty = true;
-                        Object.keys(podTargets)
-                            .filter(targetKey => podTargets[targetKey].productName === name)
-                            .map(targetKey => podTargets[targetKey].buildConfigurationList)
-                            .map(buildConfigurationListId => podConfigurationList[buildConfigurationListId])
-                            .map(buildConfigurationList => buildConfigurationList.buildConfigurations)
-                            .reduce((acc, buildConfigurations) => acc.concat(buildConfigurations), [])
-                            .map(buildConfiguration => buildConfiguration.value)
-                            .forEach(buildId => {
-                                __dirty = true;
-                                podConfigs[buildId].buildSettings.SWIFT_VERSION = swiftVersion;
-                            });
-                    }
-                });
-                if (__dirty) {
-                    fs.writeFileSync(podPbxPath, podXcodeproj.writeSync(), 'utf-8');
-                }
-            }
-        });
+        return podsjsonFile.setSwiftVersionForCocoaPodsLibraries(this.root);
     }
 
     /**
