+++ lib/prepare.js
@@ -33,6 +33,8 @@ const PlatformMunger = require('cordova-common').ConfigChanges.PlatformMunger;
 const PluginInfoProvider = require('cordova-common').PluginInfoProvider;
 const FileUpdater = require('cordova-common').FileUpdater;
 const projectFile = require('./projectFile');
+const Podfile = require('./Podfile').Podfile;
+const check_reqs = require('./check_reqs');
 
 // launch storyboard and related constants
 const IMAGESET_COMPACT_SIZE_CLASS = 'compact';
@@ -308,6 +310,22 @@ function handleBuildSettings (platformConfig, locations, infoPlist) {
     }
 
     project.write();
+
+    // If we have a Podfile, we want to update the deployment target there too
+    const podPath = path.join(locations.root, Podfile.FILENAME);
+    if (deploymentTarget && fs.existsSync(podPath)) {
+        const project_name = locations.xcodeCordovaProj.split(path.sep).pop();
+
+        const PodsJson = require('./PodsJson').PodsJson;
+        const podsjsonFile = new PodsJson(path.join(locations.root, PodsJson.FILENAME));
+        const podfileFile = new Podfile(podPath, project_name, deploymentTarget);
+        podfileFile.write();
+
+        events.emit('verbose', 'Running `pod install` (to install plugins)');
+        projectFile.purgeProjectFileCache(locations.root);
+        return podfileFile.install(check_reqs.check_cocoapods)
+            .then(() => podsjsonFile.setSwiftVersionForCocoaPodsLibraries(locations.root));
+    }
 
     return Promise.resolve();
 }
